Fix swapped name files for resource and metadata
