+++ pkg/kinds/rolebinding/rolebinding_gen.go
@@ -3,40 +3,41 @@
 // Generated by:
 //     kinds/cog_gen.go
 // Using jennies:
-//     K8MetadataJenny
+//     K8ResourcesJenny
 //
 // Run 'make gen-cue' from repository root to regenerate.
 
 package rolebinding
 
 import (
-	"time"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"github.com/grafana/grafana/pkg/kinds"
 )
 
-// Metadata defines model for Metadata.
-type Metadata struct {
-	CreatedBy         string     `json:"createdBy"`
-	CreationTimestamp time.Time  `json:"creationTimestamp"`
-	DeletionTimestamp *time.Time `json:"deletionTimestamp,omitempty"`
+// Resource is the kubernetes style representation of RoleBinding. (TODO be better)
+type K8sResource = kinds.GrafanaResource[RoleBinding, Status]
 
-	// extraFields is reserved for any fields that are pulled from the API server metadata but do not have concrete fields in the CUE metadata
-	ExtraFields     map[string]any    `json:"extraFields"`
-	Finalizers      []string          `json:"finalizers"`
-	Labels          map[string]string `json:"labels"`
-	ResourceVersion string            `json:"resourceVersion"`
-	Uid             string            `json:"uid"`
-	UpdateTimestamp time.Time         `json:"updateTimestamp"`
-	UpdatedBy       string            `json:"updatedBy"`
+// NewResource creates a new instance of the resource with a given name (UID)
+func NewK8sResource(name string, s *RoleBinding) K8sResource {
+	return K8sResource{
+		TypeMeta: v1.TypeMeta{
+			Kind:       "RoleBinding",
+			APIVersion: "v0-0-alpha",
+		},
+		ObjectMeta: v1.ObjectMeta{
+			Name:        name,
+			Annotations: make(map[string]string),
+			Labels:      make(map[string]string),
+		},
+		Spec: s,
+	}
 }
 
-// _kubeObjectMetadata is metadata found in a kubernetes object's metadata field.
-// It is not exhaustive and only includes fields which may be relevant to a kind's implementation,
-// As it is also intended to be generic enough to function with any API Server.
-type KubeObjectMetadata struct {
-	CreationTimestamp time.Time         `json:"creationTimestamp"`
-	DeletionTimestamp *time.Time        `json:"deletionTimestamp,omitempty"`
-	Finalizers        []string          `json:"finalizers"`
-	Labels            map[string]string `json:"labels"`
-	ResourceVersion   string            `json:"resourceVersion"`
-	Uid               string            `json:"uid"`
+// Resource is the wire representation of RoleBinding.
+// It currently will soon be merged into the k8s flavor (TODO be better)
+type Resource struct {
+	Metadata Metadata    `json:"metadata"`
+	Spec     RoleBinding `json:"spec"`
+	Status   Status      `json:"status"`
 }
